requirements.txt
streamlit
openai
python-dotenv
pydantic
tabulate
ibapi
mcp